Roll back to force drizzlepac 3.6.2

+++ grizli/utils.py
@@ -8232,7 +8232,7 @@ def drizzle_from_visit(
                     if jwst_dq_flags is not None:
                         bad_bits |= get_jwst_dq_bit(jwst_dq_flags, verbose=verbose)
 
-                    dq = flt[("DQ", ext)].data & bad_bits
+                    dq = flt[("DQ", ext)].data.astype(np.int32) & bad_bits
                     dq |= bpdata.astype(dq.dtype)
 
                     # Clipping threshold for BKG extensions, global at top
